Create Dalgorithm
Implemented Dijkstra's algorithm in Java. Given a text file with the start node, the total number of nodes and edges in the first line and other nodes and neighbors with their respective edge weights in the rest of the file, the code then creates an adjacency list to find the shortest possible path from the start node to every reachable node. In the end, it outputs a file with the node it reaches, the weight of the path, and the parent node in one line for each reachable node.